[web] Keyboard support for typeahead [13/13] - Freeze users list when typeahead is visible.
Summary:
One weird edge case was deleting/adding user to chat/thread when other user had overlay displayed.
Normally overlay is rerendered with newly fetched fresh user list, but it may cause weird shifts in chosen button.
We could solve it by using some extra state to save currently chosen button and then going back to it after list of users change, but I decided to go with simpler solution: when overlay is displayed list of users is frozen until it disappears.
It may result in tries to @ users that were deleted from channel, but it doesn't seem like a huge issue.
In future we may react to it by displaying modal asking if we want to sent invite to such user (or any user that we tried to @, but who was not in out channel).

Test Plan:
Added and deleted users using another account and incognito browser.
Freezing worked fine.

Reviewers: inka, kamil, ginsu, tomek

Reviewed By: tomek

Subscribers: ashoat, atul, tomek, ginsu, kamil, inka

Differential Revision: https://phab.comm.dev/D6053