[native] modified tooltip UI to show 'Unlike' when user already likes a message
Summary:
modified the tooltip react entry to either show 👍 or `Unlike` depending on whether the user liked the message. Happy to make this look better, but I figured since this is a temporary solution that will be reverted once we go from message liking to message reactions, it wasn't worth spending a ton of time to make this perfect.

I also modified the `destructiveButtonIndex` to only have a value if the `Report` option is present. As I was making the changes above, I noticed that `Unlike` was turning red when a `viewer` was trying to unlike their own message.
{F297015}

---

Depends on D5813

Test Plan:
Please watch the demo video to see the changes I made:

{F297006}

Reviewers: atul, tomek

Reviewed By: tomek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5961