[keyserver] Introduce `fetchUserIDForEthereumAddress` and use in `siweAuthResponder`
Summary:
Introduce `fetchUserIDForEthereumAddress` and use in `siweAuthResponder` to check if user exists and to complete login process (via call to `successfulLogInQueries`) if so.

---

Depends on D6059

Test Plan: Will be tested implicitly in future diffs. Will test thoroughly (lots of breakpoints to step through and ensure correctness) once account creation stuff is done.

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6060