[web] Check if copy operation is successful
Summary:
When a message is successfully copied, we set a state so that a label can be changed. At this point in the stack, the change won't be shown to the user because we need to update the tooltip.

Depends on D6075

Test Plan: Added console logs to check if the value is updated.

Reviewers: kamil, inka, ginsu, przemek

Reviewed By: kamil, przemek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6084