[native] Reverting D5999 and D6000
Summary: This diff reverts D5999 and D6000 commits that include the Rust function with using of OpenSSL. As we having a problem with the [[ https://linear.app/comm/issue/ENG-2563/update-github-ci-to-use-protobuf-315 | OpenSSL and CI building in GitHub Actions ]] these diffs will be landed after the issue would be resolved.

Test Plan: CI gates are green.

Reviewers: ashoat, tomek

Reviewed By: ashoat, tomek

Subscribers: atul

Differential Revision: https://phab.comm.dev/D6089