[native] Handle errors with SIWE WebView
Summary:
More context in this comment: [ENG-2544](https://linear.app/comm/issue/ENG-2544#comment-1b9fd9fe).
The current error message isn't great, so this diff replaces it with an alert.

{F309890}

Test Plan: For testing I forced `getSIWENonceCallFailed` to be always true and checked if the alert is displayed and if it returns to the login screen when you press "OK".

Reviewers: atul, tomek, kamil

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6097