[lib/keyserver] add check to make sure users cannot send reactions from the source message of a sidebar
Summary: While working on improving message liking, came across this edge case where users were able to react to the source message of a sidebar. We shouldn't allow users to do this, so I implemented a check on the client and keyserver to ensure we catch this behavior. In regards to displaying the reactions for the source message from the original thread to the sidebar thread, this [[ https://linear.app/comm/issue/ENG-2480/add-reaction-to-sidebar-source | linear task ]] tracks this work and is something we will be addressing in the near future.

Test Plan:
Please watch the demo videos to see the changes I made:

Before:
{F301912}

After:
{F301913}

Reviewers: atul, tomek, ashoat

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6004