[native] Add CommunityDrawerContent
Summary:
Linear issue: https://linear.app/comm/issue/ENG-1881/community-navigation-drawer-on-native
Add a custom drawer content implementation to be passed to DrawerNavigator, displaying a `CommunityDrawerItemCommunity` for each community.

Test Plan:
Added this component to `AppsDirectory` and check that it allows to navigate to threads, the elements open and close as expected. Checked that when only one community is present it is open by defalut and none is open when there are more
communities. Checked that the app dooesn't crash when there are no communities.

Reviewers: tomek, atul, ginsu, kamil

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5723