[native] Add community drawer to navigation hierarchy
Summary:
Linear issue: https://linear.app/comm/issue/ENG-1881/community-navigation-drawer-on-native and https://linear.app/comm/issue/ENG-2327/change-code-around-getstatefromnavigatorroute-to-address-adding-drawer
Adding the drawer navigator between App and Tab navigators so it opens over the tabs. Some calls to `getStateFromNavigatorRoute` had to be changed as they assumed the navigation hierarchy. The dafault
state had to be updated to relfect the new hierarchy.

Test Plan:
Checked the changes to activeThread by logging the value it returns, and the changes to baseCreateActiveTabSelector by logging the value returnes in ProfileHeader depending on the navigation state.
Run the ios simulator and checked that the drawer can be opened from the TabNavigator header and the drawer behaves as expected: CommunityDrawerContent is displayed, navigation to threads works, etc.

Reviewers: tomek, atul, ginsu, kamil, bartek, ashoat

Reviewed By: tomek, ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5734