[native] Add hamburger menu button for opening community drawer
Summary:
Linear issue: https://linear.app/comm/issue/ENG-2285/add-hamburger-menu-button-to-tabs
Add a button for opening the community drawer. The button is to be visible in all tabs when there is no "back button". Inbox and Profile tabs have custom headers, and Calendar and Apps tabs didn't use to have headers, hence the difference in how it was implemented.
Settings related to SafeAreaView had to be changed, because otherwise thare was a weird space between the header and the tab content.
Icon size, colour and margin from the edge were set to the same values used for ComposeThreadButton, so that the Chat tabs header looks consistent.
{F277724}

Test Plan:
Check that in all tabs the hamburger menu button is present and pressing it navigates to the community drawer. Check that when user navigates to a screen in Chat or Profile navigators, the button is no longer present and a "back button" appears in its place.
Tested on:
Simulators:
- iPhone 13 (has a notch)
- iPhone 13 mini (has a notch)
- iPhone SE (3rd) (no notch)
- Android 12
Physical:
- iPhone 12 (has a notch) (dev and prod)
- iPhone SE (3rd) (no notch)
- Google Pixel 6 (Android 13)

Reviewers: ginsu, kamil, tomek, ashoat

Reviewed By: tomek, ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5844