[native] Add subchannnels modal
Summary:
Linear issue: https://linear.app/comm/issue/ENG-2373/subchannels-view
Adding a subchannels modal, similar to the SidebarListModal that displays sidebars of a chat.
This modal is supposed to display subchannels of a given chat. It is to be used in the community drawer.

As the designs are [[ https://linear.app/comm/issue/DES-20/designs-for-subchannels-modal#comment-9b311fc4 | still being discussed ]], this diff doesn't introduce the final styling yet. Here is what it looks like for now:
{F298790}

Test Plan:
In chat-thread-list.react in `onPressSeeMoreSidebars` exchange the navigation call for:

```
this.props.navigation.navigate<'SubchannelsListModal'>({
      name: SubchannelsListModalRouteName,
      params: { threadInfo },
    });
```

Run ios simulator, press “See more…” under a chat that has threads and see that the SubchannelsListModal appears

Reviewers: przemek, kamil, tomek

Reviewed By: kamil, tomek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5972