[native] Add subchannels button to drawer items
Summary:
Linear issue: https://linear.app/comm/issue/ENG-2284/add-links-to-subchannels-in-level-2-chats
drawer items at level 2 are not supposed to expand further, but if they do have subchannels, expanding them should reveal a button that allows to navigate to a modal, displaying the subchannels.

{F301680}

Test Plan: Ran ios simulator, checked that buttons are present where they should be present and that they navigate to the modal.

Reviewers: ginsu, bartek, tomek

Reviewed By: tomek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6002