[web] introduce useOnClickReact hook in reaction message utils
Summary:
intorduce `useOnClickReact` hook. `useOnClickReact` is the web implmentation of the similar function `onPressReact` introduced in D5780. This function handles the calling of the create reaction message endpoint and then dispatching the results of that to the reducer

---

Linear Task: [[ https://linear.app/comm/issue/ENG-2473/add-react-button-to-web-message-tooltip | ENG-2473 ]]

Depends on D6054

Test Plan:
This task will be futher tested in subsequent diffs. I was also able to like/unlike messages in `web` in my local stack using this hook and some subsequent diffs

{F306362}

Additionally, here is what the `Alert` looks like if a message fails to go through
{F308027}

Reviewers: atul, tomek, ashoat

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6055