[keyserver] Fix `siweAuthResponder` login issue
Summary:
So after setting some breakpoints within `processSuccessfulLogin`, `fetchThreadInfos`, `fetchServerThreadInfos`, and `rawThreadInfosFromServerThreadInfos` found the issue.

Pretty dumb mistake on my part when implementing `fetchUserIDForEthereumAddress(...)`. The return value should be (and is typed as) `?string`... but was actually returning `?number` which broke some of the comparison checks in `rawThreadInfosFromServerThreadInfos`. Adding the `.toString()` at the end fixes things up and now registration and login work fully as expected.

Test Plan:
1. Able to successfully register via SIWE
2. Able to successfully log in via SIWE

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6110