[landing] Revert RainbowKit patch in D6018
Summary:
This caused [ENG-2578](https://linear.app/comm/issue/ENG-2578/android-siwe-flow-not-sending-signed-message-to-comm-app). Oops!!

The original patch just fixed behavior when no wallet was installed... otherwise the deeplink to the Play Store failed. But that's not so bad of an issue... the user can figure out a way to install the app given they selected it.

Test Plan: Test end-to-end flow on a physical Android device

Reviewers: atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6111