[native] Fix android bottom navigation bar ovelapping with drawer content
Summary:
Linear issue: https://linear.app/comm/issue/ENG-2603/community-drawer-safeareaview
I  was doing some additional testing on physical devices, and I noticed that the androids bottom navigation bar was covering some of the drawer content if the drawer content took up all drawer height. This fixes that, and here is what it looks like now with and without the bottom bar on android:
{F313335}

{F313336}

It has been decided in  [[ https://linear.app/comm/issue/ENG-2343/style-the-navigation-drawer-on-native | ENG-2343 ]] that we don't want to use 'bottom' option of SafeAreaView for iOS devices, so iOS stayed the same.

Test Plan:
Tested on two android physical devices, iOS simulator, iOS physical device. Checked that the drawer content doesn't overlap with androids navigation bar, and looks right when the bar is not present
(gestures enabled), and looks right on iOS.

Reviewers: tomek, bartek, ginsu

Reviewed By: tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6137