[CI] Skip irrelevant builds on desktop files changes
Summary: I've noticed that we are using `paths-ignore` option on GitHub CI, so I added `desktop/**` to CIs which shouldn't be affected by changes in the `desktop` workspace.

Test Plan: Tested on a fork if the CI still runs on changes outside of the `desktop` and doesn't run when the only changes are in the `desktop`.

Reviewers: atul, varun, tomek

Reviewed By: atul

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6144