[CI] Build macOS and Windows desktop app on github
Summary:
[ENG-2532](https://linear.app/comm/issue/ENG-2532)
Adds GitHub workflows for building the desktop app on Windows and macOS. The Windows one should be especially useful because nobody develops on a Windows machine so if it breaks we wouldn't catch it until later.
They are heavily inspired by the Android/ iOS workflows.

Test Plan: Tested on a fork but without secrets for now.

Reviewers: atul, tomek, jon

Reviewed By: atul

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6146