[landing] Rename `SIWENonceContext` to `SIWEContext`
Summary:
This `React.Context` was initially just to make the `siweNonce` included in the request header available in the `siwe.react.js` component without a bunch of prop drilling.

However, going forward we're also going to need to make `siwePrimaryIdentityPublicKey` accessible to the `siwe.react.js` component. Instead of introducing yet another `React.Context` it makes sense to turn `SIWENonceContext` into a more general `SIWEContext` that can handle both the nonce and public key.

This diff is **solely** about renaming the context in place to reduce clutter in subsequent diffs.

Test Plan: SIWE flow continue to work as expected.

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6127