[landing] Add `siwePrimaryIdentityPublicKey` to `SIWEContext`
Summary:
Add `siwePrimaryIdentityPublicKey` to `SIWEContext` so the value is available throughout `landing` without prop drilling. Specifically we'll be using `siwePrimaryIdentityPublicKey` in `siwe.react.js` where we'll need the public key so it can be included in the SIWE message that gets signed by the user's wallet.

---

Depends on D6127

Test Plan:
Able to retrieve the value of `siwePrimaryIdentityPublicKey` in `siwe.react.js`:
https://blob.sh/ef0479.png (yeah contrast isn't great, but the public key is there as expected).

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6128