[landing] Include `Device IdPubKey` in SIWE message statement
Summary:
If public key passed as header (`siwePrimaryIdentityPublicKey`), we'll include it in the SIWE message statement.

For the timebeing if a `siwePrimaryIdentityPublicKey` is NOT being passed in as header, we'll fallback to the previous SIWE message statement without the public key. Will go back and enforce existence of `siwePrimaryIdentityPublicKey` (with necessary error handling, error messages, etc) later in the stack.

---

Depends on D6131

Test Plan:
The `Device IdPubKey` is being included in SIWE message statement:

{F312699}

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6132