[keyserver] Pull `primaryIdentityPublicKey` out of SIWE statement and pass to `siweAccountCreationRequest`
Summary:
Use the utility functions introduced earlier in the stack to pull `primaryIdentityPublicKey` out of the signed SIWE message statement (if it exists).

For native clients we expect the value of `primaryIdentityPublicKey` to be a string that conforms to `primaryIdentityPublicKeyRegex` and for web clients we expect `null`.

---

Depends on D6132

Test Plan:
1. public key appears in signing request:
{F312726}

2. public key properly extracted on `keyserver`:
{F312727}

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6133