[native] Delete native/ios/.xcode.env.local on clean
Summary: Without this, subsequent runs of `yarn cleaninstall` do not overwrite this file. We want each run of `yarn cleaninstall` to update the file, so that a user can switch from a Homebrew-managed Node to a Nix-managed Node without needing to manually delete the file. (You just need to run `nix develop` and then `yarn cleaninstall`, and the file will be overwritten.)

Test Plan: I wrote a garbage path to `native/ios/.xcode.env.local`, then ran `yarn cleaninstall` and confirmed it was fixed

Reviewers: tomek, jon

Reviewed By: tomek

Subscribers: atul

Differential Revision: https://phab.comm.dev/D6170