[native] Get rid of "waow cookie" message
Summary:
After the React Native 0.70 upgrade, I noticed a statement saying something like "wow you found the file, here's a cookie 🍪". I couldn't find that anywhere on GitHub or via `git grep`, so I ran `grep -R 🍪 .` in the root in the repo and found what was going on.

The source was the `@favware/skip-dependency` package. I found [this issue](https://github.com/favware/skip-dependency/issues/1) on their GitHub, which revealed that the line got printed when the dependency being replaced by `@favware/skip-dependency` was included. It's there as a sort of canary to indicate that the replaced package is still being imported, and to warn of any possible issues that may result.

In our case, the import is occurring in `@react-navigation/devtools`, but since we're not using `useFlipper` from that package, nothing is done with the import, so we are safe.

To turn off the warning, I updated the package to a version that no longer printed the error, as advised in the linked GitHub issue above.

Test Plan: Confirm the cookie message was no longer getting printed in the React Native logs

Reviewers: tomek, bartek, atul

Reviewed By: tomek, bartek

Differential Revision: https://phab.comm.dev/D6194