[services][backup] CreateBackup 2/3 - handle non-data inputs
Summary:
This diff adds logic for processing non-data inputs (handler non-data-mode). After all of them are provided,
backup ID is generated and the put client is initialized and handler switches to "data mode", which is implemented in child diff.

Depends on D6197

Test Plan: The service builds and starts. Logic of this diff can be tested altogether with the child diff - see the description there.

Reviewers: varun, tomek, jon, max

Reviewed By: varun, tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6198