[services][backup] SendLog 2/4 - handle non-data inputs
Summary:
This diff adds logic for processing non-data inputs (handler non-data-mode). After all of them are provided,
log ID is generated and handler switches to "data mode" (waits for data chunks), which is implemented in next diff.

Depends on D6211

Test Plan: The service builds and starts. Logic of this diff can be tested altogether with the grandchild diff - see the description there.

Reviewers: varun, tomek, jon, max, michal

Reviewed By: tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6212