[services][backup] SendLog 3/4 - handle log data
Summary:
This diff implements receiving log contents data. By default, log data is going to be directly saved in database.
Each received data chunk is appended to the buffer. If the log item becomes too large to be stored in DynamoDB,
then log contents is moved to Blob service storage and further chunks are uploaded there directly.

Depends on D6212

Test Plan: The service builds and starts. Logic of this diff can be tested altogether with the child diff - see the description there.

Reviewers: varun, tomek, jon, max, michal

Reviewed By: varun, tomek, jon

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6213