[web] Conditionally display `ConnectButton` once `nonce` received from `keyserver`
Summary:
We'll request a `nonce` from the `keyserver` `siwe_nonce` endpoint once a wallet has been successfully connected.

We display an `ActivityIndicator` while we wait for the request to complete. If the request was successful, we'll display the `ConnectButton`. If not, the `ActivityIndicator` component will show an `!` to indicate error. Error message + handling will be improved later in the stack.

---

Depends on D6221

Test Plan:
Here's what it looks like:

Reviewers: ashoat, tomek