Display background notification from Java, remove headless JS task
Summary: This differential implements notification formatting and displaying in Java so that it can be removed from JavaScript. This allows us to remove headless JS task.

Test Plan: Build the app. Kill it. Send messages and rescinds. Notifications should appear in and disappear from NotificationCenter as they did prior to changes introduced in this differential.

Reviewers: tomek, atul

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6231