[keyserver] reaction messages only notifiy author of target message
Summary:
Reaction messages only notify author of target message. This is done by fetching the target message info and checking if the `creatorID` of the target message is the same as the `userID` of the user recieving the notif

---

Linear Task: [[ https://linear.app/comm/issue/ENG-2616/only-notify-the-author-of-a-target-message-of-a-new-reaction-message | ENG-2616 ]]

Test Plan:
Please watch the demo video to see the changes I made. For additional context ginsu is signed in to the simulator on the left and dan is signed into the simulator on the right

{F319757}

Reviewers: atul, tomek, ashoat

Reviewed By: tomek, ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6186