[keyserver] Script to create new community
Summary: Eventually we'll want to add this to the UI, but for now a script is better than nothing. This is necessary for [ENG-2416](https://linear.app/comm/issue/ENG-2416/create-a-new-community-on-comm-for-daily-updates-to-investors), and also something I called out in the [Table stakes features](https://www.notion.so/commapp/Table-stakes-features-ba47ededad0f408fa610cc7666a2a1c3) doc on Notion.

Test Plan: Ran the script locally and confirmed that new community appeared in the navigation drawer on native

Reviewers: atul, ginsu, tomek

Reviewed By: ginsu

Differential Revision: https://phab.comm.dev/D6234