[patch] Patch `@rainbow-me/rainbowkit` to increase width of formatted address
Summary:
Was struggling to get the `ConnectButton` to fit properly in context of `LogInForm`. I tried various things with CSS to get the `ConnectButton` to fit the width of its container, but had no luck.

Looks like the formatted address will always have 9 characters (4 leading, ellipsis, and 4 trailing) regardless of space available. This didn't look great in the context of `LogInForm` (very narrow) so I bumped it up to 19 characters (8 leading, ellipsis, 10 trailing).

Realize patching a package is a pretty "drastic" thing, but I read carefully through the docs/available props/etc and couldn't find any simpler or cleaner way to extend the width of the `ConnectButton`.

---

Depends on D6219

Test Plan:
Here's what it looks like:
{F325114}

**and here's what it looks like in other use in `landing`:**
{F325097}

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6220